requirements.txt
streamlit
openai
psycopg2-binary
requests
mysql-connector-python
streamlit-folium
sqlalchemy